fix: mask dtype 跨后端修复 + DirectML 冒烟脚本
- DirectML 后端 bool.sum() 返回 bool 导致专家损失崩溃(实测复现);
  5 个专家 compute_loss 与 joint_trainer 的 mask 统一转 float —
  跨后端正确性硬修复(CPU 路径 395 测试全绿不受影响)
- 新增 scripts/dml_smoke.py: DAFT 全模型在 Arc 140T (DirectML) 上
  3 步训练+推理冒烟, 已通过

+++ src/daft/training/joint_trainer.py
@@ -267,7 +267,7 @@ class JointTrainer:
                 routing_probs = outputs["routing_probs"]
 
                 # --- Joint loss: MSE on final signal + expert consistency ---
-                mse = ((signal - t_b) ** 2 * m_b).sum() / m_b.sum().clamp(min=1)
+                mse = ((signal - t_b) ** 2 * m_b.float()).sum() / m_b.float().sum().clamp(min=1)
 
                 # Expert consistency: lightly regularize each expert
                 expert_reg = 0.0
@@ -303,7 +303,7 @@ class JointTrainer:
                 with torch.no_grad():
                     outputs = self.model(s_b, layer_outputs, mode="val")
                     signal = outputs["signal"]
-                    mse = ((signal - t_b) ** 2 * m_b).sum() / m_b.sum().clamp(min=1)
+                    mse = ((signal - t_b) ** 2 * m_b.float()).sum() / m_b.float().sum().clamp(min=1)
                     total_loss += mse.item()
 
                 if return_predictions:
